% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundsGamma.R
\name{boundsGamma}
\alias{boundsGamma}
\title{Upper limits for overall mutation rate}
\usage{
boundsGamma(alleles, afreq, range)
}
\arguments{
\item{alleles}{A character vector with allele labels.}

\item{afreq}{A numeric vector of allele frequencies.}

\item{range}{A positive number.}
}
\value{
Bounds gamma for mutation matrix to be (i) well defined and (ii) regular.
}
\description{
Upper limits for overall mutation rate
}
\examples{
alleles = 1:3; afreq = c(0.2, 0.3,  0.5);  range = 0.1
bounds = boundsGamma(alleles, afreq ,range)
R1 = stepwiseReversible(alleles, afreq, rate = bounds[1], range)
isRegular(R1, afreq)
R2 = stepwiseReversible(alleles, afreq, rate = bounds[2], range)
isRegular(R2, afreq)
                        
}
\author{
Thore Egeland.
}
