% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isMutationModel}
\alias{isMutationModel}
\alias{isMutationMatrix}
\title{Test for mutation matrix/model}
\usage{
isMutationModel(x)

isMutationMatrix(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
TRUE or FALSE
}
\description{
Test for mutation matrix/model
}
\examples{

mat = mutationMatrix("equal", alleles = 1:2, rate = 0.1)
isMutationMatrix(mat)

isMutationModel(mat) # FALSE (not a complete model)

mod = mutationModel(mat)
isMutationModel(mod)

}
