% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationMatrix.R
\name{mutationMatrix}
\alias{mutationMatrix}
\title{Mutation matrix}
\usage{
mutationMatrix(model = c("custom", "equal", "proportional", "random"),
  matrix = NULL, alleles = NULL, afreq = NULL, rate = NULL,
  seed = NULL)
}
\arguments{
\item{model}{A string: either "custom", "equal", "proportional" or "random"}

\item{matrix}{When \code{model} is "custom", this must be a square matrix with
nonnegative real entries and row sums equal to 1}

\item{alleles}{A character vector (or coercible to character) with allele
labels. Required in all models, except "custom" if \code{matrix} has dimnames}

\item{afreq}{A numeric vector of allele frequencies. Required in model
"proportional"}

\item{rate}{A number between 0 and 1. Required in models "equal" and
"proportional"}

\item{seed}{A single number. Optional parameter in the "random" model, passed
on to \code{set.seed()}}
}
\value{
A square matrix with entries in \code{[0, 1]}, with the allele labels as
both colnames and rownames.
}
\description{
Construct mutation matrices for pedigree likelihood computations.
}
\details{
Descriptions of the models:
\itemize{
\item "custom" : Allows any mutation matrix to be provided by the user, in the \code{matrix}
parameter
\item "equal" :  All mutations equally likely; probability \eqn{1-rate} of no
mutation
\item "proportional" : Mutation probabilities are proportional to the target
allele frequencies
\item "random" : This produces a matrix of random numbers, where each row is
normalised so that it sums to 1
}
}
\examples{
mutationMatrix(alleles = 1:3, model = "equal", rate = 0.05)

}
