% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedmut-package.R
\docType{package}
\name{pedmut-package}
\alias{pedmut}
\alias{pedmut-package}
\title{pedmut: Mutation Models for Pedigree Likelihood Computations}
\description{
A collection of functions for modelling mutations in pedigrees with marker data, as used e.g. in likelihood computations with microsatellite data. Implemented models include equal, proportional and stepwise models, as well as random models for experimental work, and custom models allowing the user to apply any valid mutation matrix. Allele lumping is done following the lumpability criteria of Kemeny and Snell (1976), ISBN:0387901922.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/pedmut}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

Other contributors:
\itemize{
  \item Thore Egeland \email{thore.egeland@nmbu.no} (\href{https://orcid.org/0000-0002-3465-8885}{ORCID}) [contributor]
}

}
\keyword{internal}
