% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedmut-package.R
\docType{package}
\name{pedmut}
\alias{pedmut}
\title{pedmut: Mutation Models for Pedigree Likelihood Computations}
\description{
A collection of functions for modelling mutations in pedigrees with marker
data, as used e.g. in likelihood computations with microsatellite data.
Implemented models include proportional and stepwise models, as well as
random models for experimental work, and custom models allowing the user to
apply any valid mutation matrix. Allele lumping is done following the
lumpability criteria of Kemeny and Snell (1976), ISBN:0387901922.
}
