% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseReversible.R
\name{stepwiseReversible}
\alias{stepwiseReversible}
\title{Reversible stepwise mutation model}
\usage{
stepwiseReversible(alleles, afreq, rate, range)
}
\arguments{
\item{alleles}{A vector of integer integers.}

\item{afreq}{A numeric vector of allele frequencies.}

\item{rate}{A numeric mutation rate.}

\item{range}{A positive number.}
}
\value{
A reversible stepwise mutation model with expected mutation rate
equal input rate.
}
\description{
A reversible stepwise mutation model is created following the approach of
Dawid et al. (2002).
}
\details{
The function \code{\link[=maxRate]{maxRate()}} checks if it is possible to construct
the mutation matrix.
}
\examples{
stepwiseReversible(alleles = 1:3, afreq = c(0.2, 0.3,  0.5), rate = 0.001, range = 0.1)

# Model not well defined:
\dontrun{
stepwiseReversible(alleles = 1:3, afreq = c(0.2, 0.3,  0.5), rate = 0.7, range = 0.1)
}
}
\author{
Thore Egeland.
}
