% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxRate.R
\name{maxRate}
\alias{maxRate}
\title{Upper limits for overall mutation rate for the stepwise reversible model.}
\usage{
maxRate(alleles, afreq, range)
}
\arguments{
\item{alleles}{A character vector with allele labels.}

\item{afreq}{A numeric vector of allele frequencies.}

\item{range}{A positive number.}
}
\value{
Bounds on overall mutation rate for mutation matrix to be (i) well defined and (ii) regular.
}
\description{
Upper limits for overall mutation rate for the stepwise reversible model.
}
\details{
For the stepwise reversible model, the mutation rate \eqn{r_{i,j},\,  i\neq j}
is proportional to the overall mutation rate \eqn{\lambda} for given values of the range, the allele
frequency \eqn{p_i} and n, the number of alleles. Hence, we can determine bounds UW
and UB so that the model is well defined if \eqn{\lambda \leq UW} and bounded,
i.e., \eqn{r_{i,j} \leq p_j,\, i\neq j}, if \eqn{\lambda \leq UB}, The bounds UW and UB are computed.
}
\examples{
alleles = 1:3; afreq = c(0.2, 0.3,  0.5);  range = 0.1
bounds = maxRate(alleles, afreq ,range)
R1 = stepwiseReversible(alleles, afreq, rate = bounds[1], range)
isBounded(R1, afreq)
R2 = stepwiseReversible(alleles, afreq, rate = bounds[2], range)
isBounded(R2, afreq)

}
\author{
Thore Egeland.
}
