% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationMatrix.R
\name{mutationModel}
\alias{mutationModel}
\title{Mutation models}
\usage{
mutationModel(female, male = female, ...)
}
\arguments{
\item{female, male}{Either objects of class "mutationMatrix", or a character
string. In the latter case, it is passed on to \code{\link[=mutationMatrix]{mutationMatrix()}} together
with any arguments in \code{...}. If \code{male} is not explicitly given, it defaults
to be the same as \code{female}.}

\item{...}{Further arguments to \code{\link[=mutationMatrix]{mutationMatrix()}}, e.g. \code{alleles}.}
}
\description{
A mutation model is a list of two mutation matrices, named "female" and
"male".
}
